requirements.txt
aiosmtplib==4.0.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt<4.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
passlib>=1.7.4
pillow==11.3.0
pluggy==1.6.0
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
qrcode==8.2
reportlab==4.4.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
